+++ com/github/dmitrypopina/javaschool/lesson1/Ex3/Converter.java
@@ -2,5 +2,5 @@ package com.github.dmitrypopina.javaschool.lesson1.Ex3;
 
 public abstract class Converter {
     protected String listConverters = "KF";
-    abstract double convertTo(float degree);
+    abstract double convertTo(double degree);
 }
